remove redundant castings

+++ src/lib/sensor_calibration/Accelerometer.cpp
@@ -158,7 +158,7 @@ void Accelerometer::ParametersUpdate()
 		return;
 	}
 
-	_calibration_index = static_cast<int8_t>(FindCurrentCalibrationIndex(SensorString(), _device_id));
+	_calibration_index = FindCurrentCalibrationIndex(SensorString(), _device_id);
 
 	if (_calibration_index == -1) {
 		// no saved calibration available
